+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.1.3
+ * Add support for onlyRealtyIds in project call
  * Add support for project state, url, sale start, completion date and is reference in project calls
  * Add support for plain rent fields without additional costs (rent net, rent gross, rent vat type, rent vat value, rent vat input and rent vat) in realty calls
  * Add support for allProjectRealties setting